Fix a bug where starting a playground would change current directory
When starting a new playground, before executing the playground
initialistaion command, the current directory would be changed to the
snippet directory. This would have the side-effect of setting the
current directory of whatever buffer the user was in previously, which
could cause unexpected behaviour with other commands.

Instead of changing directory, locally bind default-directory so that
the initialisation command will be executed from the snippet directory
without an side effects.