fix: go-playground-upload not working in playground snippet buffers
This fixes the condition in the go-playground-upload function, so it
runs go-play-buffer within go-playground snippet file buffers only.
Previously it would run it in any buffer, except snippet buffers.

I also removed the cursor movement and insertion of the result from
go-play-upload, as default behavior of go-play-buffer and go-play-region
is to place the URL on the killring, and can be customized setting the
go-play-browse-function variable.